审核通过假期扣减单位

+++ src/main/java/com/xiangshangban/att_simple/service/ApproverServiceImpl.java
@@ -320,12 +320,12 @@ public class ApproverServiceImpl implements ApproverService {
 		//年假
 		if(leaveType.equals("0")){
 			
-			if(Integer.parseInt(leaveDay)%8==0){
-				leaveDay = String.valueOf(Integer.parseInt(leaveDay)/8+0.0);
-			}else if(Integer.parseInt(leaveDay)%8<=4){
-				leaveDay = String.valueOf(Integer.parseInt(leaveDay)/8+0.5);
-			}else if(Integer.parseInt(leaveDay)%8>4){
-				leaveDay = String.valueOf(Integer.parseInt(leaveDay)/8+1);
+			if(Double.parseDouble(leaveDay)%8==0){
+				leaveDay = String.valueOf(Double.parseDouble(leaveDay)/8+0.0);
+			}else if(Double.parseDouble(leaveDay)%8<=4){
+				leaveDay = String.valueOf(Double.parseDouble(leaveDay)/8+0.5);
+			}else if(Double.parseDouble(leaveDay)%8>4){
+				leaveDay = String.valueOf(Double.parseDouble(leaveDay)/8+1);
 			}
 			
 			year = String.valueOf(Integer.parseInt(year)-1);
